card condtiotionals figured out
